Squashed 'external/workload-automation/' changes from 832ed797e..2f231b5ce
2f231b5ce fw/target: detect module variations in TargetInfo
75878e2f2 uiauto/build_scripts: Update to use python3
023cb88ab templates/setup: Update package setup template to specify python3
d27443deb workloads/rt_app: Update to use python3
1a15f5c76 Geekbench: Adding support for Version 5 of Geekbench Corporate
d3af4e751 setup.py: Update pandas version restrictions
73b0b0d70 readthedocs: Add ReadtheDocs configuration
bb18a1a51 travis: Remove tests for python2.7
062be6d54 output_processors/postgresql: Don't reconnect if not initialised
c1e095be5 output_processors/postgresql: Ensure still connected to the database
eeebd010b output_processors/postgresql: Group database connection operations
e387e3d9b Update to remove Python2 as supported version.
6042fa374 fw/version: Version Bump
050329a5e fw/version: Update version for WA and required devlib
d9e7aa9af Dockerfile: Update to use the latest versions of WA and devlib
125cd3bb4 docs/changes: Changelog for v3.1.4
75ea78ea4 docs/faq: Fix formatting
12bb21045 instruments/SysfsExtractor: Add extracted directories as artifacts
4bb1f4988 fw/DatabaseOutput: Only attempt to extract config if avaliable
0ff6b4842 fw/DatabaseOutput: Fix the retrieval of job level artifacts
98b787e32 fw/DatabaseOutput: Enabled retrieving of directory artifacts
e91543666 commands/postgres: Upgrade the data base schema to v1.3
68e1806c0 output_processors/postgresql: Add support for new directory Artifacts
f19ebb79e output_processors/postgresql: Add missing `system_id` field
c950f5ec8 utils/postgres: Fix formatting
6aaa28781 fw/Artifact: Allows adding directories as artifacts
d87025ad3 output_processors/postgres: Fix empty iterable
ac5819da8 output_processors/postgres: Fix incorrect dict keys
31e08a647 instruments/interrupts: Add `interrupt` files as artifacts
47769cf28 Add a workload for Motionmark tests
d8601880a setup.py: Add README as package description
0efc9b9cc setup.py: Clean up extra dependencies
501d3048a requirements: Add initial version
c4daccd80 README: Update installation instruction to match documentation.
db944629f setup.py: Update classifiers
564738a2a workloads/monoperf: Fix typos
c092128e9 workloads: Sets `requires_network` attribute for workloads
463840d2b docs/faq: Add question about non UTF-8 environments.
43633ab36 extras/Dockerfile: Ensure we are using utf-8 in our docker container
a6f0ab31e fw/entrypoint: Add check for system default encoding
72fd5b513 setup.py: Set maximum package version for python2.7 support
766bb4da1 fw/uiauto: Allow specifying landscape and portrait orientation
a5f052135 utils/types: Fix typos
3435c36b9 fw/workload: Improve version matching and error propergation
bd252a647 fw/workload: Introduce max / min versions for apks
f46851a3b utils/types: Add `version_tuple`
891023444 fw/workload: Don't override the package manager for ApkRevent workloads
1108c5701 workloads: Update to better utilize `cleanup_assets` and `uninstall`
f5d1a9e94 fw/workload: Add the `uninstall` parameter to the base `workload` class
959106d61 fw/workload: Update description of `cleanup_assests` parameter
0aea3abca workloads: Add support for UIBench Jank Tests
24ccc024f framework.workload: `am instrument` APK manager
42ab81103 workloads/lmbench: Fix missing run method declaration

git-subtree-dir: external/workload-automation
git-subtree-split: 2f231b5ce54a8b6239a58b89e189a90836c1470b

+++ wa/workloads/rt_app/__init__.py
@@ -130,8 +130,8 @@ class RtApp(Workload):
                   '''),
         Parameter('cpus', kind=cpu_mask, default=0, aliases=['taskset_mask'],
                   description='Constrain execution to specific CPUs.'),
-        Parameter('uninstall_on_exit', kind=bool, default=False,
-                  description="""
+        Parameter('uninstall', aliases=['uninstall_on_exit'], kind=bool, default=False,
+                  override=True, description="""
                   If set to ``True``, rt-app binary will be uninstalled from the device
                   at the end of the run.
                   """),
@@ -213,9 +213,10 @@ class RtApp(Workload):
 
     @once
     def finalize(self, context):
-        if self.uninstall_on_exit:
+        if self.uninstall:
             self.target.uninstall(self.target_binary)
-        self.target.execute('rm -rf {}'.format(self.target_working_directory))
+        if self.cleanup_assets:
+            self.target.execute('rm -rf {}'.format(self.target_working_directory))
 
     def _deploy_rt_app_binary_if_necessary(self):
         # called from initialize() so gets invoked once per run
@@ -259,9 +260,9 @@ class RtApp(Workload):
         # use workgen dry run option to generate a use case
         # file with proper JSON grammar on host first
         try:
-            check_output('python {} -d -o {} {}'.format(self.workgen_script,
-                                                        output_file,
-                                                        user_file),
+            check_output('python3 {} -d -o {} {}'.format(self.workgen_script,
+                                                         output_file,
+                                                         user_file),
                          shell=True)
         except CalledProcessError as e:
             message = 'Could not generate config using workgen, got "{}"'
